# internal/stoker/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/steamapps": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "List known Steamapps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Continue token",
                        "name": "continue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Object"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/steamapps/{appID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get the details for a specific Steamapp ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Steamapp ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch password",
                        "name": "betapassword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Steamapp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create or update the details of a specific Steamapp ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Steamapp ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch password",
                        "name": "betapassword",
                        "in": "query"
                    },
                    {
                        "description": "Steamapp detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SteamappDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Steamapp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create or update the details of a specific Steamapp ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Steamapp ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch password",
                        "name": "betapassword",
                        "in": "query"
                    },
                    {
                        "description": "Steamapp detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SteamappDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Steamapp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/steamapps/{appID}/{branch}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get the details for a specific Steamapp ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Steamapp ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch (default public)",
                        "name": "branch",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch password",
                        "name": "betapassword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Steamapp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create or update the details of a specific Steamapp ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Steamapp ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch (default public)",
                        "name": "branch",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch password",
                        "name": "betapassword",
                        "in": "query"
                    },
                    {
                        "description": "Steamapp detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SteamappDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Steamapp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create or update the details of a specific Steamapp ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Steamapp ID",
                        "name": "appID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch (default public)",
                        "name": "branch",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Steamapp branch password",
                        "name": "betapassword",
                        "in": "query"
                    },
                    {
                        "description": "Steamapp detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SteamappDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Steamapp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "Steamapp": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "integer"
                },
                "apt_packages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "base_image": {
                    "type": "string"
                },
                "cmd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "date_created": {
                    "type": "string"
                },
                "entrypoint": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "execs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "icon_url": {
                    "type": "string"
                },
                "launch_type": {
                    "type": "string"
                },
                "locked": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "platform_type": {
                    "type": "string"
                }
            }
        },
        "SteamappDetail": {
            "type": "object",
            "properties": {
                "apt_packages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "base_image": {
                    "type": "string"
                },
                "cmd": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "entrypoint": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "execs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "launch_type": {
                    "type": "string"
                },
                "platform_type": {
                    "type": "string"
                }
            }
        }
    }
}